explore service added to keep explore tool lean
