+++ rust/crates/sift_mcp/src/service/mod.rs
@@ -1,6 +1,7 @@
 pub mod assets;
 pub mod channels;
 pub mod data;
+pub mod explore;
 pub mod ingest;
 pub mod reports;
 pub mod rules;
